src/requirements_pythonanywhere.txt
python-telegram-bot==20.8
vk-api==11.9.9
aiosqlite==0.19.0
python-dotenv==1.0.1
flask==3.0.2
gunicorn==23.0.0